refactor: prepare for different types of background tasks
