+++ zeebe/exporters/camunda-exporter/src/main/java/io/camunda/exporter/tasks/ReschedulingTask.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.exporter.archiver;
+package io.camunda.exporter.tasks;
 
 import io.camunda.zeebe.util.ExponentialBackoff;
 import java.util.concurrent.CompletableFuture;
@@ -14,8 +14,8 @@ import java.util.concurrent.TimeUnit;
 import org.slf4j.Logger;
 
 public final class ReschedulingTask implements Runnable {
-  private final ArchiverJob job;
-  private final int batchSize;
+  private final BackgroundTask task;
+  private final int minimumWorkCount;
   private final ScheduledExecutorService executor;
   private final Logger logger;
   private final ExponentialBackoff idleStrategy;
@@ -25,13 +25,13 @@ public final class ReschedulingTask implements Runnable {
   private long errorDelayMs;
 
   public ReschedulingTask(
-      final ArchiverJob job,
-      final int batchSize,
+      final BackgroundTask task,
+      final int minimumWorkCount,
       final long delayBetweenRunsMs,
       final ScheduledExecutorService executor,
       final Logger logger) {
-    this.job = job;
-    this.batchSize = batchSize;
+    this.task = task;
+    this.minimumWorkCount = minimumWorkCount;
     this.executor = executor;
     this.logger = logger;
 
@@ -41,41 +41,43 @@ public final class ReschedulingTask implements Runnable {
 
   @Override
   public void run() {
-    var batchArchived = job.archiveNextBatch();
+    var result = task.execute();
     // while we could always expect this to return a non-null result, we don't necessarily want to
     // stop, and more importantly, we want to make it transparent that something went wrong
-    if (batchArchived == null) {
+    if (result == null) {
       logger.warn(
-          "Expected to archive a batch asynchronously, but no result returned for job {}; rescheduling anyway",
-          job);
-      batchArchived = CompletableFuture.completedFuture(0);
+          "Expected to perform a background task, but no result returned for job {}; rescheduling anyway",
+          task);
+      result = CompletableFuture.completedFuture(0);
     }
 
-    batchArchived
-        .thenApplyAsync(this::onBatchArchived, executor)
-        .exceptionallyAsync(this::onArchivingError, executor)
-        .thenAcceptAsync(this::rescheduleJob, executor);
+    result
+        .thenApplyAsync(this::onWorkPerformed, executor)
+        .exceptionallyAsync(this::onError, executor)
+        .thenAcceptAsync(this::reschedule, executor);
   }
 
-  private long onBatchArchived(final int count) {
+  private long onWorkPerformed(final int count) {
     errorDelayMs = 0;
 
-    // if we worked on as much as the batch size, then there's probably even more work to
-    // be done, so use the minimum delay between runs; otherwise, backoff from the last
+    // if we worked on less than the minimum expected work count, then there's probably even more
+    // work to be done, so use the minimum delay between runs; otherwise, backoff from the last
     // known delay
-    delayMs = count >= batchSize ? idleStrategy.applyAsLong(0) : idleStrategy.applyAsLong(delayMs);
+    delayMs =
+        count >= minimumWorkCount ? idleStrategy.applyAsLong(0) : idleStrategy.applyAsLong(delayMs);
     return delayMs;
   }
 
-  private long onArchivingError(final Throwable error) {
+  private long onError(final Throwable error) {
     errorDelayMs = errorStrategy.applyAsLong(errorDelayMs);
 
-    logger.error("Error occurred while archiving data; operation will be retried", error);
+    logger.error(
+        "Error occurred while performing a background task; operation will be retried", error);
     return errorDelayMs;
   }
 
-  private void rescheduleJob(final long delay) {
-    logger.trace("Rescheduling archiving job {} in {}ms", job, delay);
+  private void reschedule(final long delay) {
+    logger.trace("Rescheduling task {} in {}ms", task, delay);
     executor.schedule(this, delay, TimeUnit.MILLISECONDS);
   }
 }
