+++ zeebe/exporters/camunda-exporter/src/test/java/io/camunda/exporter/tasks/BackgroundTaskManagerTest.java
@@ -5,12 +5,12 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.exporter.archiver;
+package io.camunda.exporter.tasks;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import io.camunda.exporter.archiver.ArchiverRepository.NoopArchiverRepository;
+import io.camunda.exporter.tasks.archiver.ArchiverRepository.NoopArchiverRepository;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
@@ -19,15 +19,16 @@ import org.junit.jupiter.api.Test;
 import org.slf4j.LoggerFactory;
 
 @AutoCloseResources
-final class ArchiverTest {
+final class BackgroundTaskManagerTest {
   @AutoCloseResource // ensures we always reap the thread no matter what
   private final ScheduledThreadPoolExecutor executor = new ScheduledThreadPoolExecutor(1);
 
   @Nested
   final class CloseTest {
     private final CloseableRepository repository = new CloseableRepository();
-    private final Archiver archiver =
-        new Archiver(1, repository, LoggerFactory.getLogger(ArchiverTest.class), executor);
+    private final BackgroundTaskManager archiver =
+        new BackgroundTaskManager(
+            1, repository, LoggerFactory.getLogger(BackgroundTaskManagerTest.class), executor);
 
     @Test
     void shouldCloseExecutorOnClose() {
